// apps/web/src/app/core/api/generated/models/organization-service-out.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */


/**
 * Estado de un servicio para una organización concreta.
 */
export interface OrganizationServiceOut {
  enabled: boolean;
  etiqueta: string;
  global_enabled: boolean;
  overridden_off: boolean;
  service_key: string;
}